create entities

+++ src/main/java/com/p1thong/springsecurity/repository/UserRepository.java
@@ -1,14 +1,15 @@
 package com.p1thong.springsecurity.repository;
 
-import com.p1thong.springsecurity.entity.Role;
-import com.p1thong.springsecurity.entity.User;
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
+import com.p1thong.springsecurity.entity.Role;
+import com.p1thong.springsecurity.entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
 
     User findByRole(Role role);
